Added minimiser base class. Cleaning up and documenting
